[be/refactor] 가격 관련 데이터 자료형 변경 (Integer -> Long)

+++ backend/auction/src/main/java/com/dokidoki/auction/domain/entity/LeaderboardEntity.java
@@ -23,12 +23,12 @@ public class LeaderboardEntity {
     private MemberEntity memberEntity;
 
     @Column(name = "bid_price")
-    private Integer bidPrice;
+    private Long bidPrice;
 
     @Column(name = "bid_time")
     private LocalDateTime bidTime;
 
-    public static LeaderboardEntity createLeaderboard(Long auctionId, MemberEntity memberEntity, Integer bidPrice, LocalDateTime bidTime) {
+    public static LeaderboardEntity createLeaderboard(Long auctionId, MemberEntity memberEntity, Long bidPrice, LocalDateTime bidTime) {
         LeaderboardEntity leaderboardEntity = new LeaderboardEntity();
         leaderboardEntity.auctionId = auctionId;
         leaderboardEntity.memberEntity = memberEntity;
