+++ backend/auction/src/main/java/com/dokidoki/auction/dto/request/AuctionUpdateReq.java
@@ -13,7 +13,7 @@ public class AuctionUpdateReq {
 
     private String description;
 
-    private Integer priceSize;
+    private Long priceSize;
 
     private String meetingPlace;
 }
